read commit description
- webhook: added url in title
- cli: change image domain to your own (set in config)
- if the configuration file cannot be open, ask the user whether to overwrite the file instead of overwriting it without asking
- logger: max_level_debug in debug binary

+++ src/config/toml.rs
@@ -1,11 +1,12 @@
 use super::Config;
 
 use anyhow::Error;
+use colored::Colorize;
 use dirs::config_dir;
 use log::warn;
 use std::{
     fs::{create_dir_all, read_to_string, File},
-    io::Write as _,
+    io::{Write as _, self},
     path::Path,
 };
 use toml::from_str as toml_from_str;
@@ -13,25 +14,39 @@ use toml::from_str as toml_from_str;
 const CONFIG_DIR: &str = "/imgurs/config.toml";
 
 pub fn parse() -> Config {
-    toml().unwrap_or_else(|e| {
-        warn!("Parse toml config: {e}! Creating config file...");
+    toml().unwrap_or_else(|err| {
+        let mut stdout = std::io::stdout();
 
-        let default_config = include_str!(concat!("../../config.toml"));
+        write!(stdout, "{}", "The configuration file could not be opened. Do you want to create/overwrite with DEFAULT values? (Y/n): ".yellow()).unwrap();
+        stdout.flush().unwrap();
 
-        let sys_config_dir = config_dir().expect("find config dir");
-        let config_dir = format!("{}{CONFIG_DIR}", sys_config_dir.to_string_lossy());
-        let config_path = Path::new(&config_dir);
+        let mut value = String::new();
+        io::stdin()
+            .read_line(&mut value)
+            .expect("failed to read line");
 
-        let prefix = config_path.parent().unwrap();
-        create_dir_all(prefix).expect("create config dir");
+        if value.to_lowercase() != "n\n" {
+            warn!("Parse toml config: {err}! Creating config file...");
 
-        let mut file_ref = File::create(config_path).expect("create failed");
+            let default_config = include_str!(concat!("../../config.toml"));
 
-        file_ref
-            .write_all(default_config.as_bytes())
-            .expect("failed write default config to file");
+            let sys_config_dir = config_dir().expect("find config dir");
+            let config_dir = format!("{}{CONFIG_DIR}", sys_config_dir.to_string_lossy());
+            let config_path = Path::new(&config_dir);
 
-        toml().expect("parse toml config")
+            let prefix = config_path.parent().unwrap();
+            create_dir_all(prefix).expect("create config dir");
+
+            let mut file_ref = File::create(config_path).expect("create failed");
+
+            file_ref
+                .write_all(default_config.as_bytes())
+                .expect("failed write default config to file");
+
+            toml().expect("parse toml config")
+        } else {
+            panic!("New config creation cancelled!")
+        }
     })
 }
